refactor: nuxt example
