+++ sandbox/nuxt/pages/index.tsx
@@ -1,4 +1,4 @@
-import Input from '../components/Input.vue'
+import VModel from '../components/VModel.vue'
 import CompositionApiDemo from '../components/CompositionApiDemo.vue'
 import SetupDemo from '../components/SetupDemo.vue'
 import Styled from '../components/Styled.vue'
@@ -11,7 +11,7 @@ export default defineComponent(() => {
       <SetupDemo />
       <Styled>styled</Styled>
       <br />
-      <Input />
+      <VModel />
     </>
   )
 })
