+++ sandbox/nuxt/panda.config.ts
@@ -6,5 +6,5 @@ export default defineConfig({
   exclude: [],
   outdir: 'styled-system',
   jsxFramework: 'vue',
-  jsxStyleProps: 'minimal',
+  jsxStyleProps: 'all',
 })
